move shell level test to new file
